Correct implicit and explicit property inheritance
Setting up a config where a child class inherits its parent's  properties
through class inheritance did not work.

The logic in TypesGenerator handled implicit inheritance based on
whether the parent's properties key was not set or not an array, but did
not account for empty arrays.

Also added tests for these implicit and explicit cases.