fix: Check if key is set to prevent warning
When the properitesMap array gets iterated to generate the fields, it happens that the uri of the typeInHierarchy is not a key of the said array; this triggers a PHP Warning:  Undefined array key. Seems to happen only for custom classes definition though. This does not prevent the command to run, this is really not critical.