feat: add resolveTypes and allTypes parameters
The allTypes parameter can be set on a specific vocabulary to generate
all its types on top of the explicit types you have provided.
If the allTypes parameter is enabled globally, you can also use it to
disable it on a specific vocabulary (but you can explicitly define
types for this vocabulary).
If resolveTypes is enabled, if a type is present in a vocabulary but not
explicitly imported or if the vocabulary is not totally imported
(allTypes), it will be generated.